add debug version of driver1 script
 - uses a constant source term to produce spatially varying but constant
   in time potential and electric fields
 - disables redundant field solves in the time stepping loop
 - uses a non-physical source term

This version of the problem can be compared to a known solution and
exercises the field gather part of the PIC algorithm.  It does not
exercise the charge scatter part such that if the numerical solution
agrees with the analytic solution to this point, it is a good indicator
that any issues likely are related to the charge scatter step.